requirements.txt
fastapi
uvicorn[standard]
requests
python-docx
pdfminer.six
beautifulsoup4
scikit-learn
